Add Doxygen support
Modifications to pepper the generated .h files with Doxygen comments. Also
a template Doxyfile is provided but the INPUT/FILE_PATTERNS needs to be
edited to name just ONE of the .h files that is generated (this does not,
yet, attempt to make collections of .h files that can all co-exist in one
set of documentation)

Invoke the Doxygen stuff uding -d on the command line to the Python
script.